helpers and straglers
