+++ eq-author-api/schema/resolvers/utils/folders.js
@@ -1,7 +1,7 @@
 const { find, flatMap, some } = require("lodash");
 const { v4: uuidv4 } = require("uuid");
 const { getSections, getSectionById } = require("./sections");
-
+const { createCalculatedSummary, createQuestionPage } = require("./pages");
 const getFolders = (ctx) => flatMap(getSections(ctx), ({ folders }) => folders);
 
 const getFoldersBySectionId = (ctx, id) => getSectionById(ctx, id).folders;
