+++ eq-author-api/schema/resolvers/utils/pages.js
@@ -1,4 +1,5 @@
-const { get, find, flatMap, some } = require("lodash");
+const { get, compact, find, flatMap, some } = require("lodash");
+const { v4: uuidv4 } = require("uuid");
 const { getFolders, getFolderById } = require("./folders");
 const { getSectionById } = require("./sections");
 
@@ -33,6 +34,66 @@ const getPageByValidationId = (ctx, validationId) =>
     (page) => page.totalValidation && page.totalValidation.id === validationId
   );
 
+const getPosition = (position, comparator) =>
+  typeof position === "number" ? position : comparator.length;
+
+const getMovePosition = (section, pageId, position) => {
+  if (!section.folders) {
+    throw new Error("Section doesn't have a folder");
+  }
+
+  let pointer = 0;
+  let positionMap = {};
+
+  for (let i = 0; i < section.folders.length; i++) {
+    for (let j = 0; j < section.folders[i].pages.length; j++) {
+      const page = section.folders[i].pages[j];
+      if (page.id === pageId) {
+        positionMap.previous = {
+          folderIndex: i,
+          pageIndex: j,
+          page,
+        };
+      }
+      if (pointer === position) {
+        positionMap.next = { folderIndex: i };
+      }
+      pointer++;
+    }
+  }
+
+  const { previous, next } = positionMap;
+  return { previous, next };
+};
+
+const getConfirmations = (ctx) =>
+  compact(flatMap(getPages(ctx), (page) => page.confirmation));
+
+const getConfirmationById = (ctx, id) => find(getConfirmations(ctx), { id });
+
+const createQuestionPage = (input = {}) => ({
+  id: uuidv4(),
+  pageType: "QuestionPage",
+  title: "",
+  description: "",
+  descriptionEnabled: false,
+  guidanceEnabled: false,
+  definitionEnabled: false,
+  additionalInfoEnabled: false,
+  answers: [],
+  routing: null,
+  alias: null,
+  ...input,
+});
+
+const createCalculatedSummary = (input = {}) => ({
+  id: uuidv4(),
+  title: "",
+  pageType: "CalculatedSummaryPage",
+  summaryAnswers: [],
+  ...input,
+});
+
 module.exports = {
   getPages,
   getPagesBySectionId,
@@ -42,4 +103,10 @@ module.exports = {
   getPageByConfirmationId,
   getPageByValidationId,
   getPageByAnswerId,
+  getMovePosition,
+  getPosition,
+  getConfirmations,
+  getConfirmationById,
+  createQuestionPage,
+  createCalculatedSummary,
 };
