+++ eq-author-api/src/validation/customKeywords/validateLeftHandSide.js
@@ -8,7 +8,9 @@ const {
 const createValidationError = require("../createValidationError");
 
 const { getPath } = require("../utils");
-const { getAbsolutePositionById } = require("../../../schema/resolvers/utils");
+const {
+  getAbsolutePositionById,
+} = require("../../../schema/resolvers/utils/helpers");
 
 module.exports = function (ajv) {
   ajv.addKeyword("validateLeftHandSide", {
